Use Darwinic means to get a kernel version
`syscall.Uname` doesn't exist in BSD/Darwin; construct a kernel
version string by getting the OS release via `syscall.Sysctl(...)`
instead.

NB this only matters for `go install cmd/fluxd`, since the container
image binaries are always built for Linux.
