+++ cmd/fluxd/checkpoint_darwin.go
@@ -1,0 +1,13 @@
+package main
+
+import (
+	"syscall"
+)
+
+func getKernelVersion() string {
+	v, err := syscall.Sysctl("kern.osrelease")
+	if err != nil {
+		panic(err)
+	}
+	return "darwin-" + v
+}
